Relocation of CodeMirrorEditorProps
